Stop applying workaround for libusb < 1.0.9
Librtlsdr has a workaround for libusb versions that lack 
libusb_handle_events_timeout_completed, which was added in version 1.0.9 
(released 2012-04-02). The workaround is always applied unless the 
HAVE_LIBUSB_HANDLE_EVENTS_TIMEOUT_COMPLETED macro is set, but the cmake 
code that sets this macro was removed in 
849f8efca42b659bf7e8fe17156ee0aa67b47233. As a result, the workaround is 
now always applied. This results in an extra 1-second delay whenever a 
GNU Radio flowgraph containing an RTL-SDR block is stopped, which makes 
operations like switching between demodulators in Gqrx annoyingly slow.

To solve this problem, I've simply removed the workaround, as it should 
no longer be needed.

I wonder if perhaps the workaround recently applied in 
2659e2df31e592d74d6dd264a4f5ce242c6369c8 might stem from the same bug.